# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meticulous API",
    "description": "REST API for the Meticulous CI/CD platform",
    "contact": {
      "name": "Meticulous Team"
    },
    "license": {
      "name": "UNLICENSED"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/v1/tokens": {
      "get": {
        "tags": [
          "tokens"
        ],
        "summary": "List tokens for the current user.",
        "operationId": "list_tokens",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tokens",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "tokens"
        ],
        "summary": "Create a new API token.",
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTokenResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/tokens/{id}": {
      "delete": {
        "tags": [
          "tokens"
        ],
        "summary": "Revoke an API token.",
        "operationId": "revoke_token",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Token ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token revoked"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Token not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTokenRequest": {
        "type": "object",
        "description": "Create token request body.",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description of the token's purpose."
          },
          "expires_in_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Expiration in days (None = never expires)."
          },
          "name": {
            "type": "string",
            "description": "Display name for the token."
          },
          "project_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Project IDs this token can access (None = all projects)."
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Permission scopes for this token."
          }
        }
      },
      "CreateTokenResponseBody": {
        "type": "object",
        "description": "Create token response (includes the plain token).",
        "required": [
          "token",
          "plain_token"
        ],
        "properties": {
          "plain_token": {
            "type": "string",
            "description": "The plain token value (only shown once)."
          },
          "token": {
            "$ref": "#/components/schemas/TokenResponse",
            "description": "The created token metadata."
          }
        }
      },
...
      "TokenResponse": {
        "type": "object",
        "description": "Token response (public, without hash).",
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "project_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}